dashboard/requirements.txt
streamlit
pandas
matplotlib
Pillow
numpy
plotly
